fix: correct type conflict

+++ 14-binary_tree_balance.c
@@ -17,7 +17,7 @@ int binary_tree_balance(const binary_tree_t *tree)
 	left_height = binary_tree_height(tree->left);
 	right_height = binary_tree_height(tree->right);
 
-	return (left_height - right_height);
+	return ((int)left_height - (int)right_height);
 }
 
 /**
@@ -26,9 +26,9 @@ int binary_tree_balance(const binary_tree_t *tree)
  *
  * Return: Height of the tree, 0 if tree is NULL
  */
-int binary_tree_height(const binary_tree_t *tree)
+size_t binary_tree_height(const binary_tree_t *tree)
 {
-	int left_height, right_height;
+	size_t left_height, right_height;
 
 	if (!tree)
 		return (0);
